Update build system for VS2022/Boost1.90/Python3 — Explore2Game builds and runs
- CMakePresets.json + vcpkg.json: bake in Qt5.15, OpenCL, vcpkg, boost-python (Python 3.12)
- C++17 mode (/std:c++17) required for LogMessage copy elision, auto_ptr removal
- Boost 1.90 API: io_service→io_context, from_string→make_address, is_complete→is_absolute
- sse_utils.hpp: _THROW0→noexcept, MMX intrinsics→SSE2 (_mm_cvtepi32_pd/_mm_cvtpd_epi32)
- boost_python_wrapper.h: Python 2→3 compat (#define initVice PyInit_Vice)
- stl_aligned_allocator.h: rebind now preserves Alignment template parameter
- rendergl.cpp/rendergl_2.h: comparator operator() marked const
- spatialnoise.inl, perlin.inl: remove default args from definitions (C++17 strict)
- Profiler.h: std::auto_ptr→unique_ptr
- Vice: link Boost::python + Boost::filesystem; add boost/container_hash for hash_combine
- hpalib: add CMakeLists.txt and full source (C library for high-precision floats)
- Explore2.cpp: drop stale 4th arg to initOpenGL, cast GLDEBUGPROC, explicit variant ctor

+++ Explore2Engine/chunklod.cpp
@@ -364,7 +364,7 @@ ChunkLOD::Errors ChunkLOD::calculate_errors(const Vector3f& localCameraPos, Erro
 
 	float offs = 0.0001f;
 	Errors childErrors;
-	childErrors.show = std::min(std::pow(_childError, 2) * shortestD_2_recip + offs, errors.show);
+	childErrors.show = std::min((float)(std::pow(_childError, 2) * shortestD_2_recip + offs), errors.show);
 	childErrors.create = std::max((errors.show + childErrors.show) * 0.5f, childErrors.show);
 	return childErrors;
 }
